+++ integration-tests/js-compute/fixtures/tla/src/hello-world.js
@@ -1,5 +1,7 @@
 import { routes } from './routes';
 
-routes.set('/hello-world', () => {
-  return new Response('hello world');
+routes.set('/hello-world', (evt) => {
+  console.log(evt.request);
+  const res = new Response('hello world');
+  return res;
 });