+++ runtime/fastly/builtins/fastly.cpp
@@ -49,7 +49,8 @@ const JSErrorFormatString *FastlyGetErrorMessage(void *userRef, unsigned errorNu
 JS::PersistentRooted<JSObject *> Fastly::env;
 JS::PersistentRooted<JSObject *> Fastly::baseURL;
 JS::PersistentRooted<JSString *> Fastly::defaultBackend;
-bool Fastly::allowDynamicBackends = false;
+bool allowDynamicBackendsCalled = false;
+bool Fastly::allowDynamicBackends = true;
 
 bool Fastly::dump(JSContext *cx, unsigned argc, JS::Value *vp) {
   JS::CallArgs args = CallArgsFromVp(argc, vp);
@@ -217,7 +218,11 @@ bool Fastly::createFanoutHandoff(JSContext *cx, unsigned argc, JS::Value *vp) {
   }
 
   auto backend_value = args.get(1);
-  auto backend_chars = core::encode(cx, backend_value);
+  JS::RootedString backend_str(cx, JS::ToString(cx, backend_value));
+  if (!backend_str) {
+    return false;
+  }
+  auto backend_chars = core::encode(cx, backend_str);
   if (!backend_chars) {
     return false;
   }
@@ -233,9 +238,10 @@ bool Fastly::createFanoutHandoff(JSContext *cx, unsigned argc, JS::Value *vp) {
 
   bool is_upstream = true;
   bool is_grip_upgrade = true;
+
   JS::RootedObject response(cx, Response::create(cx, response_instance, response_handle.unwrap(),
                                                  body_handle.unwrap(), is_upstream, is_grip_upgrade,
-                                                 std::move(backend_chars.ptr)));
+                                                 backend_str));
   if (!response) {
     return false;
   }
@@ -373,6 +379,9 @@ bool Fastly::defaultBackend_set(JSContext *cx, unsigned argc, JS::Value *vp) {
     return false;
 
   defaultBackend = backend;
+  if (!allowDynamicBackendsCalled) {
+    allowDynamicBackends = false;
+  }
   args.rval().setUndefined();
   return true;
 }
@@ -395,6 +404,7 @@ bool Fastly::allowDynamicBackends_set(JSContext *cx, unsigned argc, JS::Value *v
   } else {
     allowDynamicBackends = JS::ToBoolean(set_value);
   }
+  allowDynamicBackendsCalled = true;
   args.rval().setUndefined();
   return true;
 }