+++ test-d/env.test-d.ts
@@ -1,5 +1,5 @@
 /// <reference path="../types/env.d.ts" />
-import { env } from "fastly:env";
+import { env } from 'fastly:env';
 import { expectType } from 'tsd';
 
-expectType<(key: string) => string>(env)
+expectType<(key: string) => string>(env);