+++ test-d/experimental.test-d.ts
@@ -1,9 +1,22 @@
 /// <reference path="../types/experimental.d.ts" />
-import { setBaseURL, setDefaultBackend, enableDebugLogging, includeBytes, allowDynamicBackends } from "fastly:experimental";
+import {
+  setBaseURL,
+  setDefaultBackend,
+  enableDebugLogging,
+  includeBytes,
+  allowDynamicBackends,
+} from 'fastly:experimental';
 import { expectType } from 'tsd';
 
-expectType<(path: string) => Uint8Array>(includeBytes)
-expectType<(enabled: boolean) => void>(enableDebugLogging)
-expectType<(base: URL | null | undefined) => void>(setBaseURL)
-expectType<(backend: string) => void>(setDefaultBackend)
-expectType<{ (enabled: boolean): void; (defaultConfig: { connectTimeout?: number | undefined; firstByteTimeout?: number | undefined; betweenBytesTimeout?: number | undefined; }): void; }>(allowDynamicBackends)
+expectType<(path: string) => Uint8Array>(includeBytes);
+expectType<(enabled: boolean) => void>(enableDebugLogging);
+expectType<(base: URL | null | undefined) => void>(setBaseURL);
+expectType<(backend: string) => void>(setDefaultBackend);
+expectType<{
+  (enabled: boolean): void;
+  (defaultConfig: {
+    connectTimeout?: number | undefined;
+    firstByteTimeout?: number | undefined;
+    betweenBytesTimeout?: number | undefined;
+  }): void;
+}>(allowDynamicBackends);