+++ types/config-store.d.ts
@@ -1,5 +1,4 @@
-
-declare module "fastly:config-store" {
+declare module 'fastly:config-store' {
   /**
    * Class for accessing [Fastly Edge Dictionaries](https://docs.fastly.com/en/guides/about-edge-dictionaries).
    *
@@ -8,16 +7,15 @@ declare module "fastly:config-store" {
   class ConfigStore {
     /**
      * Creates a new ConfigStore object
-     * 
+     *
      * @throws {Error} Throws an `Error` if no Config Store exists with the provided name
      */
     constructor(name: string);
     /**
      * Get a value for a key in the config-store. If the provided key does not exist in the Config Store then this returns `null`.
-     * 
+     *
      * @throws {TypeError} Throws an `TypeError` if the provided key is longer than 256 characters.
      */
     get(key: string): string | null;
   }
-
 }