+++ types/env.d.ts
@@ -1,17 +1,17 @@
-declare module "fastly:env" {
+declare module 'fastly:env' {
   /**
    * Function to get the value for the provided environment variable name.
    *
    * For a list of available environment variables, see the [Fastly Developer Hub for Compute Environment Variables](https://developer.fastly.com/reference/compute/ecp-env/)
-   * 
+   *
    * **Note**: The environment variables can only be retrieved when processing requests, not during build-time initialization.
    *
    * @param name The name of the environment variable
-   * 
+   *
    * @example
    * <script async defer src="https://fiddle.fastly.dev/embed.js"></script>
    * In this example we log to stdout the environment variables `FASTLY_HOSTNAME` and `FASTLY_TRACE_ID`.
-   * 
+   *
    * <script type="application/json+fiddle">
    * {
    *   "type": "javascript",
@@ -43,16 +43,16 @@ declare module "fastly:env" {
    * ```js
    * /// <reference types="@fastly/js-compute" />
    * import { env } from "fastly:env";
-   * 
+   *
    * function app(event) {
    *   console.log("FASTLY_HOSTNAME:", env("FASTLY_HOSTNAME"));
    *   console.log("FASTLY_TRACE_ID:", env("FASTLY_TRACE_ID"));
-   * 
+   *
    *   return new Response("", {
    *     status: 200
    *   });
    * }
-   * 
+   *
    * addEventListener("fetch", event => event.respondWith(app(event)));
    * ```
    * </noscript>