+++ types/kv-store.d.ts
@@ -1,34 +1,32 @@
-/// <reference path="globals.d.ts" />
-
-declare module "fastly:kv-store" {
+declare module 'fastly:kv-store' {
   /**
    * Class for accessing a [Fastly KV-store](https://developer.fastly.com/reference/api/kv-store/).
    *
    * A kv store is a persistent, globally consistent key-value store.
    *
    * **Note**: Can only be used when processing requests, not during build-time initialization.
-   * 
+   *
    * @example
    * <script async defer src="https://fiddle.fastly.dev/embed.js"></script>
    * In this example we connect to an KV Store named `'files'` and save an entry to the store under the key `'hello'` and then read back the value and return it to the client.
-   * 
+   *
    * ```js
    * /// <reference types="@fastly/js-compute" />
-   * 
+   *
    * import { KVStore } from "fastly:kv-store";
-   * 
+   *
    * async function app(event) {
    *   const files = new KVStore('files')
-   * 
+   *
    *   await files.put('hello', 'world')
-   * 
+   *
    *   const entry = await files.get('hello')
-   * 
+   *
    *   return new Response(await entry.text())
    * }
-   * 
+   *
    * addEventListener("fetch", (event) => event.respondWith(app(event)))
-   * 
+   *
    * ```
    */
   export class KVStore {